Move DatabaseItem to separate file
